+++ amazeing-bin/src/mode/visualize.rs
@@ -19,7 +19,7 @@ async fn looper() {
 async fn main() {
     let (screen_width, screen_height) = DRAW_CTX.read().unwrap().screen_size();
 
-    set_window_size(screen_width as u32, screen_height as u32 + 30);
+    set_window_size(screen_width, screen_height + 30);
 
     looper().await
 }
